Update caching logic and fix store value conversions

+++ plugins/ethereum-adapter/src/main/java/org/polypheny/db/adapter/ethereum/EthereumDataSource.java
@@ -524,7 +524,7 @@ public class EthereumDataSource extends DataSource {
                 return PolyType.VARCHAR;
             case "int": // 8 to 256...
             case "uint256":
-                return PolyType.BIGINT;
+                return PolyType.DECIMAL; // todo
             default:
                 return null;
         }
